Moving MissingFieldException (fully) and MissingMethodException (partially) to shared (#18207)
