+++ src/System.Private.CoreLib/shared/System/MissingMemberException.cs
@@ -2,24 +2,13 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-/*=============================================================================
-**
-**
-**
-** Purpose: The exception class for versioning problems with DLLS.
-**
-**
-=============================================================================*/
-
-
 using System.Runtime.Serialization;
-using System.Runtime.CompilerServices;
 
 namespace System
 {
     [Serializable]
     [System.Runtime.CompilerServices.TypeForwardedFrom("mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089")]
-    public class MissingMemberException : MemberAccessException, ISerializable
+    public partial class MissingMemberException : MemberAccessException
     {
         public MissingMemberException()
             : base(SR.Arg_MissingMemberException)
@@ -39,37 +28,18 @@ namespace System
             HResult = HResults.COR_E_MISSINGMEMBER;
         }
 
-        protected MissingMemberException(SerializationInfo info, StreamingContext context) : base(info, context)
+        public MissingMemberException(string className, string memberName)
+        {
+            ClassName = className;
+            MemberName = memberName;
+        }
+
+        protected MissingMemberException(SerializationInfo info, StreamingContext context)
+            : base(info, context)
         {
             ClassName = info.GetString("MMClassName");
             MemberName = info.GetString("MMMemberName");
             Signature = (byte[])info.GetValue("MMSignature", typeof(byte[]));
-        }
-
-        public override String Message
-        {
-            get
-            {
-                if (ClassName == null)
-                {
-                    return base.Message;
-                }
-                else
-                {
-                    // do any desired fixups to classname here.
-                    return SR.Format(SR.MissingMember_Name, ClassName + "." + MemberName + (Signature != null ? " " + FormatSignature(Signature) : ""));
-                }
-            }
-        }
-
-        // Called to format signature
-        [MethodImplAttribute(MethodImplOptions.InternalCall)]
-        internal static extern String FormatSignature(byte[] signature);
-
-        public MissingMemberException(String className, String memberName)
-        {
-            ClassName = className;
-            MemberName = memberName;
         }
 
         public override void GetObjectData(SerializationInfo info, StreamingContext context)
@@ -80,12 +50,27 @@ namespace System
             info.AddValue("MMSignature", Signature, typeof(byte[]));
         }
 
+        public override string Message
+        {
+            get
+            {
+                if (ClassName == null)
+                {
+                    return base.Message;
+                }
+                else
+                {
+                    // do any desired fixups to classname here.
+                    return SR.Format(SR.MissingMember_Name, ClassName + "." + MemberName + (Signature != null ? " " + FormatSignature(Signature) : string.Empty));
+                }
+            }
+        }
 
         // If ClassName != null, GetMessage will construct on the fly using it
         // and the other variables. This allows customization of the
         // format depending on the language environment.
-        protected String ClassName;
-        protected String MemberName;
+        protected string ClassName;
+        protected string MemberName;
         protected byte[] Signature;
     }
 }
