+++ src/System.Private.CoreLib/src/System/Diagnostics/Stacktrace.cs
@@ -2,199 +2,38 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System;
 using System.Collections;
 using System.Collections.Generic;
+using System.Globalization;
+using System.Reflection;
+using System.Runtime.CompilerServices;
+using System.Security;
 using System.Text;
 using System.Threading;
-using System.Security;
-using System.IO;
-using System.Reflection;
-using System.Runtime.InteropServices;
-using System.Runtime.CompilerServices;
-using System.Globalization;
-using System.Runtime.Versioning;
 
 namespace System.Diagnostics
 {
-    // READ ME:
-    // Modifying the order or fields of this object may require other changes 
-    // to the unmanaged definition of the StackFrameHelper class, in 
-    // VM\DebugDebugger.h. The binder will catch some of these layout problems.
-    internal class StackFrameHelper
-    {
-        private Thread targetThread;
-        private int[] rgiOffset;
-        private int[] rgiILOffset;
-
-#pragma warning disable 414
-        // dynamicMethods is an array of System.Resolver objects, used to keep
-        // DynamicMethodDescs alive for the lifetime of StackFrameHelper.
-        private Object dynamicMethods; // Field is not used from managed.        
-
-        private IntPtr[] rgMethodHandle;
-        private string[] rgAssemblyPath;
-        private IntPtr[] rgLoadedPeAddress;
-        private int[] rgiLoadedPeSize;
-        private IntPtr[] rgInMemoryPdbAddress;
-        private int[] rgiInMemoryPdbSize;
-        // if rgiMethodToken[i] == 0, then don't attempt to get the portable PDB source/info
-        private int[] rgiMethodToken;
-        private string[] rgFilename;
-        private int[] rgiLineNumber;
-        private int[] rgiColumnNumber;
-        private bool[] rgiLastFrameFromForeignExceptionStackTrace;
-        private int iFrameCount;
-#pragma warning restore 414
-
-        private delegate void GetSourceLineInfoDelegate(string assemblyPath, IntPtr loadedPeAddress, int loadedPeSize,
-            IntPtr inMemoryPdbAddress, int inMemoryPdbSize, int methodToken, int ilOffset,
-            out string sourceFile, out int sourceLine, out int sourceColumn);
-
-        private static GetSourceLineInfoDelegate s_getSourceLineInfo = null;
-
-        [ThreadStatic]
-        private static int t_reentrancy = 0;
-
-        public StackFrameHelper(Thread target)
-        {
-            targetThread = target;
-            rgMethodHandle = null;
-            rgiMethodToken = null;
-            rgiOffset = null;
-            rgiILOffset = null;
-            rgAssemblyPath = null;
-            rgLoadedPeAddress = null;
-            rgiLoadedPeSize = null;
-            rgInMemoryPdbAddress = null;
-            rgiInMemoryPdbSize = null;
-            dynamicMethods = null;
-            rgFilename = null;
-            rgiLineNumber = null;
-            rgiColumnNumber = null;
-
-            rgiLastFrameFromForeignExceptionStackTrace = null;
-
-            // 0 means capture all frames.  For StackTraces from an Exception, the EE always
-            // captures all frames.  For other uses of StackTraces, we can abort stack walking after
-            // some limit if we want to by setting this to a non-zero value.  In Whidbey this was 
-            // hard-coded to 512, but some customers complained.  There shouldn't be any need to limit
-            // this as memory/CPU is no longer allocated up front.  If there is some reason to provide a
-            // limit in the future, then we should expose it in the managed API so applications can 
-            // override it.
-            iFrameCount = 0;
-        }
-
-        //
-        // Initializes the stack trace helper. If fNeedFileInfo is true, initializes rgFilename, 
-        // rgiLineNumber and rgiColumnNumber fields using the portable PDB reader if not already
-        // done by GetStackFramesInternal (on Windows for old PDB format).
-        //
-        internal void InitializeSourceInfo(int iSkip, bool fNeedFileInfo, Exception exception)
-        {
-            StackTrace.GetStackFramesInternal(this, iSkip, fNeedFileInfo, exception);
-
-            if (!fNeedFileInfo)
-                return;
-
-            // Check if this function is being reentered because of an exception in the code below
-            if (t_reentrancy > 0)
-                return;
-
-            t_reentrancy++;
-            try
-            {
-                if (s_getSourceLineInfo == null)
-                {
-                    Type symbolsType = Type.GetType(
-                        "System.Diagnostics.StackTraceSymbols, System.Diagnostics.StackTrace, Version=4.0.1.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a",
-                        throwOnError: false);
-
-                    if (symbolsType == null)
-                    {
-                        return;
-                    }
-
-                    MethodInfo symbolsMethodInfo = symbolsType.GetMethod("GetSourceLineInfo", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance);
-                    if (symbolsMethodInfo == null)
-                    {
-                        return;
-                    }
-
-                    // Create an instance of System.Diagnostics.Stacktrace.Symbols
-                    object target = Activator.CreateInstance(symbolsType);
-
-                    // Create an instance delegate for the GetSourceLineInfo method
-                    GetSourceLineInfoDelegate getSourceLineInfo = (GetSourceLineInfoDelegate)symbolsMethodInfo.CreateDelegate(typeof(GetSourceLineInfoDelegate), target);
-
-                    // We could race with another thread. It doesn't matter if we win or lose, the losing instance will be GC'ed and all threads including this one will
-                    // use the winning instance
-                    Interlocked.CompareExchange(ref s_getSourceLineInfo, getSourceLineInfo, null);
-                }
-
-                for (int index = 0; index < iFrameCount; index++)
-                {
-                    // If there was some reason not to try get the symbols from the portable PDB reader like the module was
-                    // ENC or the source/line info was already retrieved, the method token is 0.
-                    if (rgiMethodToken[index] != 0)
-                    {
-                        s_getSourceLineInfo(rgAssemblyPath[index], rgLoadedPeAddress[index], rgiLoadedPeSize[index],
-                            rgInMemoryPdbAddress[index], rgiInMemoryPdbSize[index], rgiMethodToken[index],
-                            rgiILOffset[index], out rgFilename[index], out rgiLineNumber[index], out rgiColumnNumber[index]);
-                    }
-                }
-            }
-            catch
-            {
-            }
-            finally
-            {
-                t_reentrancy--;
-            }
-        }
-
-        public virtual MethodBase GetMethodBase(int i)
-        {
-            // There may be a better way to do this.
-            // we got RuntimeMethodHandles here and we need to go to MethodBase
-            // but we don't know whether the reflection info has been initialized
-            // or not. So we call GetMethods and GetConstructors on the type
-            // and then we fetch the proper MethodBase!!
-            IntPtr mh = rgMethodHandle[i];
-
-            if (mh == IntPtr.Zero)
-                return null;
-
-            IRuntimeMethodInfo mhReal = RuntimeMethodHandle.GetTypicalMethodDefinition(new RuntimeMethodInfoStub(mh, this));
-
-            return RuntimeType.GetMethodBase(mhReal);
-        }
-
-        public virtual int GetOffset(int i) { return rgiOffset[i]; }
-        public virtual int GetILOffset(int i) { return rgiILOffset[i]; }
-        public virtual string GetFilename(int i) { return rgFilename == null ? null : rgFilename[i]; }
-        public virtual int GetLineNumber(int i) { return rgiLineNumber == null ? 0 : rgiLineNumber[i]; }
-        public virtual int GetColumnNumber(int i) { return rgiColumnNumber == null ? 0 : rgiColumnNumber[i]; }
-
-        public virtual bool IsLastFrameFromForeignExceptionStackTrace(int i)
-        {
-            return (rgiLastFrameFromForeignExceptionStackTrace == null) ? false : rgiLastFrameFromForeignExceptionStackTrace[i];
-        }
-
-        public virtual int GetNumberOfFrames() { return iFrameCount; }
-    }
-
-
-    // Class which represents a description of a stack trace
-    // There is no good reason for the methods of this class to be virtual.  
+    /// <summary>
+    /// Class which represents a description of a stack trace
+    /// There is no good reason for the methods of this class to be virtual.
+    /// In order to ensure trusted code can trust the data it gets from a
+    /// StackTrace, we use an InheritanceDemand to prevent partially-trusted
+    /// subclasses.
+    /// </summary>
     public class StackTrace
     {
-        private StackFrame[] frames;
         private int m_iNumOfFrames;
         public const int METHODS_TO_SKIP = 0;
         private int m_iMethodsToSkip;
+        
+        /// <summary>
+        /// Stack frames comprising this stack trace.
+        /// </summary>
+        private StackFrame[] _stackFrames;
 
-        // Constructs a stack trace from the current location.
+        /// <summary>
+        /// Constructs a stack trace from the current location.
+        /// </summary>
         public StackTrace()
         {
             m_iNumOfFrames = 0;
@@ -202,8 +41,9 @@ namespace System.Diagnostics
             CaptureStackTrace(METHODS_TO_SKIP, false, null, null);
         }
 
-        // Constructs a stack trace from the current location.
-        //
+        /// <summary>
+        /// Constructs a stack trace from the current location.
+        /// </summary>
         public StackTrace(bool fNeedFileInfo)
         {
             m_iNumOfFrames = 0;
@@ -211,9 +51,10 @@ namespace System.Diagnostics
             CaptureStackTrace(METHODS_TO_SKIP, fNeedFileInfo, null, null);
         }
 
-        // Constructs a stack trace from the current location, in a caller's
-        // frame
-        //
+        /// <summary>
+        /// Constructs a stack trace from the current location, in a caller's
+        /// frame
+        /// </summary>
         public StackTrace(int skipFrames)
         {
             if (skipFrames < 0)
@@ -226,9 +67,10 @@ namespace System.Diagnostics
             CaptureStackTrace(skipFrames + METHODS_TO_SKIP, false, null, null);
         }
 
-        // Constructs a stack trace from the current location, in a caller's
-        // frame
-        //
+        /// <summary>
+        /// Constructs a stack trace from the current location, in a caller's
+        /// frame
+        /// </summary>
         public StackTrace(int skipFrames, bool fNeedFileInfo)
         {
             if (skipFrames < 0)
@@ -241,8 +83,9 @@ namespace System.Diagnostics
             CaptureStackTrace(skipFrames + METHODS_TO_SKIP, fNeedFileInfo, null, null);
         }
 
-
-        // Constructs a stack trace from the current location.
+        /// <summary>
+        /// Constructs a stack trace from the current location.
+        /// </summary>
         public StackTrace(Exception e)
         {
             if (e == null)
@@ -253,8 +96,9 @@ namespace System.Diagnostics
             CaptureStackTrace(METHODS_TO_SKIP, false, null, e);
         }
 
-        // Constructs a stack trace from the current location.
-        //
+        /// <summary>
+        /// Constructs a stack trace from the current location.
+        /// </summary>
         public StackTrace(Exception e, bool fNeedFileInfo)
         {
             if (e == null)
@@ -265,9 +109,10 @@ namespace System.Diagnostics
             CaptureStackTrace(METHODS_TO_SKIP, fNeedFileInfo, null, e);
         }
 
-        // Constructs a stack trace from the current location, in a caller's
-        // frame
-        //
+        /// <summary>
+        /// Constructs a stack trace from the current location, in a caller's
+        /// frame
+        /// </summary>
         public StackTrace(Exception e, int skipFrames)
         {
             if (e == null)
@@ -283,9 +128,10 @@ namespace System.Diagnostics
             CaptureStackTrace(skipFrames + METHODS_TO_SKIP, false, null, e);
         }
 
-        // Constructs a stack trace from the current location, in a caller's
-        // frame
-        //
+        /// <summary>
+        /// Constructs a stack trace from the current location, in a caller's
+        /// frame
+        /// </summary>
         public StackTrace(Exception e, int skipFrames, bool fNeedFileInfo)
         {
             if (e == null)
@@ -301,14 +147,13 @@ namespace System.Diagnostics
             CaptureStackTrace(skipFrames + METHODS_TO_SKIP, fNeedFileInfo, null, e);
         }
 
-
-        // Constructs a "fake" stack trace, just containing a single frame.  
-        // Does not have the overhead of a full stack trace.
-        //
+        /// <summary>
+        /// Constructs a "fake" stack trace, just containing a single frame.
+        /// Does not have the overhead of a full stack trace.
+        /// </summary>
         public StackTrace(StackFrame frame)
         {
-            frames = new StackFrame[1];
-            frames[0] = frame;
+            _stackFrames = new StackFrame[] { frame };
             m_iMethodsToSkip = 0;
             m_iNumOfFrames = 1;
         }
@@ -345,9 +190,10 @@ namespace System.Diagnostics
             return iRetVal;
         }
 
-        // Retrieves an object with stack trace information encoded.
-        // It leaves out the first "iSkip" lines of the stacktrace.
-        //
+        /// <summary>
+        /// Retrieves an object with stack trace information encoded.
+        /// It leaves out the first "iSkip" lines of the stacktrace.
+        /// </summary>
         private void CaptureStackTrace(int iSkip, bool fNeedFileInfo, Thread targetThread, Exception e)
         {
             m_iMethodsToSkip += iSkip;
@@ -363,7 +209,7 @@ namespace System.Diagnostics
 
             if (m_iNumOfFrames != 0)
             {
-                frames = new StackFrame[m_iNumOfFrames];
+                _stackFrames = new StackFrame[m_iNumOfFrames];
 
                 for (int i = 0; i < m_iNumOfFrames; i++)
                 {
@@ -384,7 +230,7 @@ namespace System.Diagnostics
                         sfTemp.SetColumnNumber(StackF.GetColumnNumber(i));
                     }
 
-                    frames[i] = sfTemp;
+                    _stackFrames[i] = sfTemp;
                 }
 
                 // CalculateFramesToSkip skips all frames in the System.Diagnostics namespace,
@@ -402,56 +248,61 @@ namespace System.Diagnostics
             // In case this is the same object being re-used, set frames to null
             else
             {
-                frames = null;
+                _stackFrames = null;
             }
         }
 
-        // Property to get the number of frames in the stack trace
-        //
+        /// <summary>
+        /// Property to get the number of frames in the stack trace
+        /// </summary>
         public virtual int FrameCount
         {
             get { return m_iNumOfFrames; }
         }
 
-
-        // Returns a given stack frame.  Stack frames are numbered starting at
-        // zero, which is the last stack frame pushed.
-        //
+        /// <summary>
+        /// Returns a given stack frame.  Stack frames are numbered starting at
+        /// zero, which is the last stack frame pushed.
+        /// </summary>
         public virtual StackFrame GetFrame(int index)
         {
-            if ((frames != null) && (index < m_iNumOfFrames) && (index >= 0))
-                return frames[index + m_iMethodsToSkip];
+            if ((_stackFrames != null) && (index < m_iNumOfFrames) && (index >= 0))
+                return _stackFrames[index + m_iMethodsToSkip];
 
             return null;
         }
 
-        // Returns an array of all stack frames for this stacktrace.
-        // The array is ordered and sized such that GetFrames()[i] == GetFrame(i)
-        // The nth element of this array is the same as GetFrame(n). 
-        // The length of the array is the same as FrameCount.
-        // 
+        /// <summary>
+        /// Returns an array of all stack frames for this stacktrace.
+        /// The array is ordered and sized such that GetFrames()[i] == GetFrame(i)
+        /// The nth element of this array is the same as GetFrame(n).
+        /// The length of the array is the same as FrameCount.
+        /// </summary>
         public virtual StackFrame[] GetFrames()
         {
-            if (frames == null || m_iNumOfFrames <= 0)
+            if (_stackFrames == null || m_iNumOfFrames <= 0)
                 return null;
 
             // We have to return a subset of the array. Unfortunately this
             // means we have to allocate a new array and copy over.
             StackFrame[] array = new StackFrame[m_iNumOfFrames];
-            Array.Copy(frames, m_iMethodsToSkip, array, 0, m_iNumOfFrames);
+            Array.Copy(_stackFrames, m_iMethodsToSkip, array, 0, m_iNumOfFrames);
             return array;
         }
 
-        // Builds a readable representation of the stack trace
-        //
+        /// <summary>
+        /// Builds a readable representation of the stack trace
+        /// </summary>
         public override string ToString()
         {
             // Include a trailing newline for backwards compatibility
             return ToString(TraceFormat.TrailingNewLine);
         }
 
-        // TraceFormat is Used to specify options for how the 
-        // string-representation of a StackTrace should be generated.
+        /// <summary>
+        /// TraceFormat is Used to specify options for how the 
+        /// string-representation of a StackTrace should be generated.
+        /// </summary>
         internal enum TraceFormat
         {
             Normal,
@@ -459,8 +310,10 @@ namespace System.Diagnostics
             NoResourceLookup    // to prevent infinite resource recusion
         }
 
-        // Builds a readable representation of the stack trace, specifying 
-        // the format for backwards compatibility.
+        /// <summary>
+        /// Builds a readable representation of the stack trace, specifying 
+        /// the format for backwards compatibility.
+        /// </summary>
         internal string ToString(TraceFormat traceFormat)
         {
             bool displayFilenames = true;   // we'll try, but demand may fail
