+++ src/System.Private.CoreLib/src/System/MissingMemberException.CoreCLR.cs
@@ -1,0 +1,15 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Runtime.Serialization;
+using System.Runtime.CompilerServices;
+
+namespace System
+{
+    public partial class MissingMemberException : MemberAccessException, ISerializable
+    {
+        [MethodImplAttribute(MethodImplOptions.InternalCall)]
+        internal static extern string FormatSignature(byte[] signature);
+    }
+}
