+++ src/System.Private.CoreLib/src/System/ThrowHelper.cs
@@ -4,8 +4,8 @@
 
 
 // This file defines an internal class used to throw exceptions in BCL code.
-// The main purpose is to reduce code size.
-//
+// The main purpose is to reduce code size. 
+// 
 // The old way to throw an exception generates quite a lot IL code and assembly code.
 // Following is an example:
 //     C# source
@@ -17,10 +17,10 @@
 //          IL_0012:  newobj     instance void System.ArgumentNullException::.ctor(string,string)
 //          IL_0017:  throw
 //    which is 21bytes in IL.
-//
+// 
 // So we want to get rid of the ldstr and call to Environment.GetResource in IL.
 // In order to do that, I created two enums: ExceptionResource, ExceptionArgument to represent the
-// argument name and resource name in a small integer. The source code will be changed to
+// argument name and resource name in a small integer. The source code will be changed to 
 //    ThrowHelper.ThrowArgumentNullException(ExceptionArgument.key, ExceptionResource.ArgumentNull_Key);
 //
 // The IL code will be 7 bytes.
@@ -29,16 +29,16 @@
 //    IL_000a:  call       void System.ThrowHelper::ThrowArgumentNullException(valuetype System.ExceptionArgument)
 //    IL_000f:  ldarg.0
 //
-// This will also reduce the Jitted code size a lot.
+// This will also reduce the Jitted code size a lot. 
 //
-// It is very important we do this for generic classes because we can easily generate the same code
-// multiple times for different instantiation.
-//
+// It is very important we do this for generic classes because we can easily generate the same code 
+// multiple times for different instantiation. 
+// 
 
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.Runtime.CompilerServices;
 using System.Runtime.Serialization;
-using System.Diagnostics;
 
 namespace System
 {
